+++ app/src/main/java/com/rodcollab/brupapp/hangman/ui/components/game/FilledWord.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.brupapp.hangman.ui.components
+package com.rodcollab.brupapp.hangman.ui.components.game
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.rodcollab.brupapp.hangman.ui.CharItem
+import com.rodcollab.brupapp.hangman.ui.model.CharItem
 
 @Composable
 fun FilledWord(modifier: Modifier, chars: List<CharItem>) {
