+++ app/src/main/java/com/rodcollab/brupapp/hangman/ui/components/game/Keyboard.kt
@@ -1,11 +1,11 @@
-package com.rodcollab.brupapp.hangman.ui.components
+package com.rodcollab.brupapp.hangman.ui.components.game
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.rodcollab.brupapp.hangman.ui.LetterModel
+import com.rodcollab.brupapp.hangman.ui.model.LetterModel
 
 @Composable
 fun KeyBoard(
