+++ app/src/main/java/com/rodcollab/brupapp/hangman/ui/components/game/ScoreHeader.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.brupapp.hangman.ui.components
+package com.rodcollab.brupapp.hangman.ui.components.game
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
