+++ app/src/main/java/com/rodcollab/brupapp/hangman/ui/model/CharItem.kt
@@ -1,10 +1,11 @@
-package com.rodcollab.brupapp.hangman.ui
+package com.rodcollab.brupapp.hangman.ui.model
 
 data class CharItem(
     val char: String,
     val guessed: Boolean,
 ) {
     val isLetter: Boolean = isLetter(char.toCharArray().first())
+    val alpha = if(guessed || !isLetter) 1.0f else 0.0f
 }
 
 fun List<Char>.toCharItem(usedLetters: List<Char>) = map { char ->
